fix: change getLikes action logic
